# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent-Native Middleware API",
    "description": "## Agent-Native Middleware API\n\nHeadless infrastructure for the B2A (Business-to-Agent) economy. This API provides four core services that give AI agents the 'hands' and 'eyes' to manipulate the digital and physical world.\n\n### Service Pillars\n\n- **IoT Protocol Bridge** (`/v1/iot`) \u2014 Secure, ACL-enforced protocol translation for IoT devices (MQTT, CoAP, Zigbee, etc.)\n- **Autonomous Product Manager** (`/v1/telemetry`) \u2014 Ingest telemetry, detect anomalies, auto-generate pull requests to fix bugs\n- **Programmatic Media Engine** (`/v1/media`) \u2014 Video ingestion, viral hook detection, reframing, captioning, and cross-platform distribution\n- **Agent Communications** (`/v1/comms`) \u2014 Agent registration, structured messaging, capability discovery, and swarm task handoffs\n\n### Authentication\n\nAll endpoints require an API key passed via the `X-API-Key` header.\n\n### For Agents\n\nThis API is designed for programmatic consumption. See `/llm.txt` for LLM-optimized documentation and `/openapi.json` for the full spec.",
    "contact": {
      "name": "Agent-Native Middleware",
      "email": "api@yourdomain.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.yourdomain.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
...
    }
  ],
  "paths": {
...
    "/v1/comms/agents": {
      "post": {
        "tags": [
          "Agent Communications"
        ],
        "summary": "Register an agent",
        "description": "Register a new agent in the communication network. The agent receives a unique ID and API key for messaging. Capabilities are used by other agents to discover and hand off tasks.",
        "operationId": "register_agent_v1_comms_agents_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRegistrationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Agent Communications"
        ],
        "summary": "List registered agents",
        "description": "List all agents in the communication network, optionally filtered by capability.",
        "operationId": "list_agents_v1_comms_agents_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "capability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter agents by capability (e.g., 'video-transcription').",
              "title": "Capability"
            },
            "description": "Filter agents by capability (e.g., 'video-transcription')."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/comms/messages": {
      "post": {
        "tags": [
          "Agent Communications"
        ],
        "summary": "Send a message to another agent",
        "description": "Send a structured message to a registered agent. Messages are routed based on priority: critical messages are delivered immediately with aggressive retry, low priority messages are batched. If the recipient has a webhook, delivery is push-based; otherwise the recipient must poll.",
        "operationId": "send_message_v1_comms_messages_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "from_agent",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "From Agent"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/comms/messages/{agent_id}/inbox": {
      "get": {
        "tags": [
          "Agent Communications"
        ],
        "summary": "Poll for messages",
        "description": "Poll an agent's inbox for pending messages. Messages are marked as delivered upon retrieval. Use this if the agent does not have a webhook configured.",
        "operationId": "poll_inbox_v1_comms_messages__agent_id__inbox_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max messages to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max messages to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/comms/messages/{agent_id}/ack/{message_id}": {
      "post": {
        "tags": [
          "Agent Communications"
        ],
        "summary": "Acknowledge a message",
        "description": "Confirm receipt and processing of a message.",
        "operationId": "acknowledge_message_v1_comms_messages__agent_id__ack__message_id__post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/comms/handoff": {
      "post": {
        "tags": [
          "Agent Communications"
        ],
        "summary": "Request a capability-based task handoff",
        "description": "Find an agent with the required capability and hand off work to it. This is the core swarm intelligence mechanism: instead of one 'God model' doing everything, tasks are routed to specialist agents. Returns the handoff message and target agent details.",
        "operationId": "request_handoff_v1_comms_handoff_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "from_agent",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "From Agent"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing API key"
          },
          "403": {
            "description": "Invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentRegistrationRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable agent name.",
            "examples": [
              "iot-monitor-agent"
            ]
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities",
            "description": "List of capabilities this agent provides. Other agents use these to discover and hand off tasks.",
            "examples": [
              [
                "iot-monitoring",
                "mqtt-bridging",
                "device-management"
              ]
            ]
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url",
            "description": "URL for push-based message delivery. If null, agent must poll.",
            "examples": [
              "https://my-agent.example.com/webhook/messages"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "capabilities"
        ],
        "title": "AgentRegistrationRequest",
        "description": "Register an agent in the communication network."
      },
      "AgentRegistrationResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "Agent-specific API key for sending/receiving messages. Store securely."
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "name",
          "capabilities",
          "api_key",
          "webhook_url",
          "status",
          "registered_at"
        ],
        "title": "AgentRegistrationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HandoffRequest": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability",
            "description": "The capability needed. System finds the best available agent.",
            "examples": [
              "video-transcription"
            ]
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context",
            "description": "Context for the handoff \u2014 what needs to be done and why.",
            "examples": [
              {
                "language": "en",
                "reason": "hooks need transcript",
                "video_id": "abc-123"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "capability",
          "context"
        ],
        "title": "HandoffRequest",
        "description": "Request a task handoff to an agent with a specific capability."
      },
      "HandoffResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Agent Id"
          },
          "target_agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Agent Name"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HandoffResponse"
      },
      "MessagePriority": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "normal",
          "low"
        ],
        "title": "MessagePriority"
      },
      "MessageResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "from_agent": {
            "type": "string",
            "title": "From Agent"
          },
          "to_agent": {
            "type": "string",
            "title": "To Agent"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "delivery_status": {
            "type": "string",
            "title": "Delivery Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "delivered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "from_agent",
          "to_agent",
          "message_type",
          "priority",
          "subject",
          "delivery_status",
          "created_at",
          "delivered_at"
        ],
        "title": "MessageResponse"
      },
      "MessageType": {
        "type": "string",
        "enum": [
          "request",
          "response",
          "event",
          "heartbeat",
          "handoff"
        ],
        "title": "MessageType",
        "description": "Standard message types for agent-to-agent communication."
      },
...
      "SendMessageRequest": {
        "properties": {
          "to_agent": {
            "type": "string",
            "title": "To Agent",
            "description": "Recipient agent ID."
          },
          "message_type": {
            "$ref": "#/components/schemas/MessageType",
            "description": "Type of message: request, response, event, heartbeat, or handoff.",
            "default": "request"
          },
          "priority": {
            "$ref": "#/components/schemas/MessagePriority",
            "description": "Delivery priority. Critical = immediate retry, Low = batched.",
            "default": "normal"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Brief subject line for the message.",
            "examples": [
              "Anomaly detected in IoT bridge"
            ]
          },
          "body": {
            "additionalProperties": true,
            "type": "object",
            "title": "Body",
            "description": "Structured message payload. Must be JSON-serializable."
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id",
            "description": "Links related messages (e.g., request/response pairs)."
          },
          "reply_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply To",
            "description": "Message ID this is replying to."
          }
        },
        "type": "object",
        "required": [
          "to_agent",
          "subject",
          "body"
        ],
        "title": "SendMessageRequest",
        "description": "Send a message to another agent."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key for agent authentication. Pass in the X-API-Key header.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}